+++ CHANGELOG.md
@@ -17,6 +17,8 @@ Note: Try to add categories to changes and link to MRs/Issues
 - [BUG] Ensure pyproject.toml symlink is created correctly inside the backed project folder
 - [NEW] Allow using `cmd` argument to pass extra parameters to `make test-node` and `make test-django`
 - [NEW] Add support for automatic deploys from Gitlab  - !228
+- [NEW] Add configurable health checks support  - !238
+- [NEW] Add support for case-insensitive emails - !238 & !221
 - [NEW] WEBAPP: Upgrade to webpack 5 - !239
 - [BUG] WEBAPP: Fix build error due to incompatibility with node-sass==6 and sass-loader versions - !239
 - [NEW] WEBAPP: Set SECURE_PROXY_SSL_HEADER for webapp too to ensure build_absolute_uri always uses correct protocol - !239